Fix CI workflow (again)
It appears that recently github has removed the dzen2 package from the runners? Though it seems to vary from
run to run whether it's present. herbstluftwm fails to launch if its missing. Unfortunately dzen2 doesn not
seem to work with cache-apt-pkgs so we have to do a full install each time now.